requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
beautifulsoup4
readability-lxml
duckduckgo_search
python-dotenv
